fix sean doubt regarding db update on thumbsUp
