+++ tree/main/extensions/speechgpt/components/ModelSelection.tsx
@@ -13,7 +13,8 @@ const ModelSelection = () => {
   })
 
   return (
-    <div className="mt-5">
+    <div className="mt-5
+    ">
       <Select
         className="mt-2"
         placeholder={model}
