+++ tree/main/extensions/speechgpt/components/SideBar.tsx
@@ -1,68 +1,130 @@
-'use client'
+"use client"
+
+
+
+import React, { useRef } from "react";
+import classNames from "classnames";
+import Link from "next/link";
+import Image from "next/image";
+import {
+  ChevronDoubleLeftIcon,
+  ChevronDoubleRightIcon,
+} from "@heroicons/react/24/outline";
 
 import NewChat from "./NewChat";
 import { signOut, useSession } from "next-auth/react";
-import {useCollection} from "react-firebase-hooks/firestore";
-import {collection, orderBy, query} from "firebase/firestore";
-import {db} from "../firebase";
+import { useCollection } from "react-firebase-hooks/firestore";
+import { collection, orderBy, query } from "firebase/firestore";
+import { db } from "../firebase";
 import ChatRow from "./ChatRow";
 import ModelSelection from "./ModelSelection";
-function SideBar() {
+
+// add NavItem prop to component prop
+type Props = {
+  collapsed: boolean;
+  setCollapsed(collapsed: boolean): void;
+  shown: boolean;
+};
+const SideBar = ({
+  collapsed,
+  shown,
+  setCollapsed,
+}: Props) => {
+  const Icon = collapsed ? ChevronDoubleRightIcon : ChevronDoubleLeftIcon;
 
 
-  const {data: session} = useSession();
+  const { data: session } = useSession();
 
   const [chats, loading, error] = useCollection(
     session && query(collection(db, "users", session?.user?.email!, "chats"),
-    orderBy("createdAt", "asc"))
+      orderBy("createdAt", "asc"))
   );
 
 
-  return <div className="flex flex-col h-screen p-2">
-    <div className="flex-1">
-      <div>
-        
-        {/* New Chat */}
-        <NewChat />
+  return (
+    <div
+      className={classNames({
+        " text-zinc-50 fixed md:static md:translate-x-0 z-20":
+          true,
+        "transition-all duration-300 ease-in-out": true,
+        "w-[300px]": !collapsed,
+        "w-16": collapsed,
+        "-translate-x-full": !shown,
+      })}
+    >
+      <div
+        className={classNames({
+          "flex flex-col justify-between h-screen sticky inset-0 w-full": true,
+        })}
+      >
+        {/* logo and collapse button */}
+        <div
+          className={classNames({
+            "flex items-center border-b border-b-white transition-none":
+              true,
+            "p-4 justify-between": !collapsed,
+            "py-4 justify-center": collapsed,
+          })}
+        >
+          {!collapsed && <div className="flex flex-col">
+            <NewChat />
 
-        <div className="hidden sm:inline">
-          <ModelSelection></ModelSelection>
+            <div className="">
+              <ModelSelection></ModelSelection>
+            </div></div>}
+          <button
+            className="grid w-10 h-10 rounded-full opacity-0 place-content-center hover:bg-white md:opacity-100"
+            onClick={() => setCollapsed(!collapsed)}
+          >
+            <Icon className="w-5 h-5" />
+          </button>
         </div>
+        <nav className="flex-grow">
+          <ul
+            className={classNames({
+              "my-2 flex flex-col gap-2 items-stretch": true,
+            })}
+          >
 
-        <div className="flex flex-col my-2 space-y-2">
-
-          {loading && (
-            <div className="text-center text-white animate-pulse">
-              <p>Loading Chats...</p>
+            {loading && (
+              <div className="text-center text-white animate-pulse">
+                <p>Loading Chats...</p>
               </div>
+            )}
+
+            {
+              chats?.docs.map((chat) => (
+                <ChatRow key={chat.id} id={chat.id} />
+              )
+              )
+            }
+          </ul>
+        </nav>
+        <div
+          className={classNames({
+            "grid place-content-stretch p-4 ": true,
+          })}
+        >
+
+          {!collapsed && (
+            <> {
+              session && (
+                // log out button
+                <a onClick={() => signOut()} className="flex items-center gap-3 px-3 py-3 text-sm text-white transition-colors duration-200 rounded-md cursor-pointer hover:bg-gray-500/10">
+                  <svg stroke="currentColor" fill="none" stroke-width="2" viewBox="0 0 24 24" stroke-linecap="round" stroke-linejoin="round" className="w-4 h-4" height="1em" width="1em" xmlns="http://www.w3.org/2000/svg">
+                    <path d="M9 21H5a2 2 0 0 1-2-2V5a2 2 0 0 1 2-2h4"></path>
+                    <polyline points="16 17 21 12 16 7"></polyline>
+                    <line x1="21" y1="12" x2="9" y2="12"></line>
+                  </svg>Log out</a>
+
+              )
+            }</>
+
           )}
-
-        {
-          chats?.docs.map((chat) => (
-            <ChatRow key={chat.id} id={chat.id} />
-          )
-        )
-      }
         </div>
-
-
-      
       </div>
     </div>
-
-    {
-      session && (
-        // log out button
-        <a onClick={() => signOut()} className="flex items-center gap-3 px-3 py-3 text-sm text-white transition-colors duration-200 rounded-md cursor-pointer hover:bg-gray-500/10">
-          <svg stroke="currentColor" fill="none" stroke-width="2" viewBox="0 0 24 24" stroke-linecap="round" stroke-linejoin="round" className="w-4 h-4" height="1em" width="1em" xmlns="http://www.w3.org/2000/svg">
-            <path d="M9 21H5a2 2 0 0 1-2-2V5a2 2 0 0 1 2-2h4"></path>
-            <polyline points="16 17 21 12 16 7"></polyline>
-            <line x1="21" y1="12" x2="9" y2="12"></line>
-          </svg>Log out</a>
-        
-      )
-    }
-  </div>; 
-}
-
+  );
+};
 export default SideBar;
+
